pyproject.toml
[project]
name = "openai-agents-sdk-practice"
version = "0.1.0"
description = "A project for practicing OpenAI Agents SDK key features."
requires-python = ">=3.9"
dependencies = [
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
    "jupyterlab>=4.0.0",
    "openai-agents>=0.2.0",
    "gradio>=4.0.0",
]

[project.optional-dependencies]
dev = []

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []